refactor(cli): slim detail pipeline review cleanup
